+++ CHANGELOG.md
@@ -9,6 +9,6 @@ All Notable changes to `Hex` will be documented in this file.
 
 ### Added
 
-- `Event` entity
+- `Reservation` entity
 - `Email` value object